Aran A/backend/requirements.txt
asgiref==3.4.1
certifi==2021.5.30
Django==3.2.6
django-cors-headers==3.8.0
django-environ==0.4.5
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
environ==1.0
Pillow==8.3.2
PyJWT==2.1.0
pytz==2021.1
sqlparse==0.4.1
validate-email==1.3
waitress==2.0.0
whitenoise==6.0.0
wincertstore==0.2
